# Support MarshalText/UnmarshalText for map keys